+++ src/main/java/vendingmachine/domain/Coin.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.domain;
 
 public enum Coin {
     COIN_500(500),
