+++ src/main/java/vendingmachine/domain/VendingMachine.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.domain;
 
 import vendingmachine.view.InputView;
